Clean up unused variables, imports in other plugins
