+++ flows/_shared_flow_utils/types.py
@@ -1,4 +1,3 @@
-import re
 from enum import Enum
 from typing import Optional, Literal
 from pydantic import BaseModel, SecretStr
@@ -56,16 +55,6 @@ class RequestType(str, Enum):
     DELETE = "delete"
 
 
-class LiquibaseAction(str, Enum):
-    UPDATE = "update"  # Create and update schema
-    UPDATECOUNT = "updateCount"  # Create schema with count
-    STATUS = "status"  # Get Version Info
-    ROLLBACK_COUNT = "rollbackCount"  # Rollback on n changesets
-    ROLLBACK_TAG = "rollback"  # Rollback on tag
-    # mark all changesets in databasechangelog table as executed
-    CHANGELOG_SYNC = "changelog-sync"
-
-
 class InternalPluginType(str, Enum):
     DATA_MANAGEMENT = "data_management_plugin"
     DATA_CHARACTERIZATION = "data_characterization_plugin"
@@ -101,12 +90,3 @@ class EntityCountDistributionType(BaseModel):
 class AuthToken(RunInput):
     token: SecretStr
     thirdpartytoken: SecretStr
-
-
-CHARACTERIZATION_DATA_MODEL = "characterization"
-OMOP_DATA_MODELS = ["omop", "omop5-4", "custom-omop-ms", "custom-omop-ms-phi"]
-CHANGESET_AVAILABLE_REGEX = re.compile(r"db/migrations/\S+")
-LB_ERROR_MESSAGE_REGEX = re.compile(r"Unexpected error running Liquibase:")
-PASSWORD_REGEX = re.compile(r"password=\S+")
-SSL_TRUST_STORE_REGEX = re.compile(
-    r"&sslTrustStore=-----BEGIN CERTIFICATE-----[a-zA-Z0-9\+\/]+-----END CERTIFICATE-----")
