+++ flows/base/data_characterization_plugin/types.py
@@ -1,10 +1,6 @@
-from enum import Enum
 from typing import Optional
 from pydantic import BaseModel
 
-FLOW_NAME = "data_characterization_plugin"
-CHARACTERIZATION_DATA_MODEL = "characterization"
-CHANGELOG_FILE = "liquibase-characterization.xml"
 
 class DCOptionsType(BaseModel):
     schemaName: str
