+++ flows/data_management/data_management_plugin/const.py
@@ -1,4 +1,3 @@
-from os import getcwd
 from re import compile
 
 from _shared_flow_utils.dao.DBDao import DBDao
@@ -12,6 +11,15 @@ DATAMODEL_CDM_VERSION = {
     "custom-omop-ms": "5.3",
     "custom-omop-ms-phi": "5.3"
 }
+
+CHANGESET_AVAILABLE_REGEX = compile(r"db/migrations/\S+")
+
+LB_ERROR_MESSAGE_REGEX = compile(r"Unexpected error running Liquibase:")
+
+PASSWORD_REGEX = compile(r"password=\S+")
+
+SSL_TRUST_STORE_REGEX = compile(
+    r"&sslTrustStore=-----BEGIN CERTIFICATE-----[a-zA-Z0-9\+\/]+-----END CERTIFICATE-----")
 
 
 def hana_to_postgres(table_name: str) -> str:
@@ -41,8 +49,3 @@ def convert_case(entity, is_lower_case):
         return entity.upper()
 
 
-CHANGESET_AVAILABLE_REGEX = compile(r"db/migrations/\S+")
-LB_ERROR_MESSAGE_REGEX = compile(r"Unexpected error running Liquibase:")
-PASSWORD_REGEX = compile(r"password=\S+")
-SSL_TRUST_STORE_REGEX = compile(
-    r"&sslTrustStore=-----BEGIN CERTIFICATE-----[a-zA-Z0-9\+\/]+-----END CERTIFICATE-----")
